storage: show the partitioning request review for the 'Use configured storage'
When the user exits the cockpit-storage, they need to see somewhere a
review of their actions. Show that as a hint item, in the scenario
description.
