+++ src/components/storage/InstallationScenario.scss
@@ -7,5 +7,9 @@
       font-weight: var(--pf-v5-global--FontWeight--bold);
       margin-inline-end: var(--pf-v5-global--spacer--xs);
     }
+
+    .installation-method-scenario-review {
+        color: var(--pf-v5-c-radio__description--Color);
+    }
   }
 }
